backend/api_foodgram/requirements.txt
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3

requests==2.26.0
asgiref==3.2.10
python-dotenv==0.20.0
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.9.5
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1


django-templated-mail==1.1.1
isort==5.10.1
python3-openid==3.2.0
requests==2.26.0
requests-oauthlib==1.3.1
django-extensions==2.2.9


Pillow==8.3.1